requirements.txt
torch>=1.9.0
numpy>=1.19.5
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
prometheus-client>=0.11.0
